requirement.txt
streamlit>=1.28.0
speechrecognition>=3.10.0
pydub>=0.25.1
pronouncing>=0.2.0
gtts>=2.3.2
playsound>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.13.0
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11
